Refactor Work in progress
Move code out of main
Stop using global variables for LoomoSDK
Separate ROS activity
Reimplement viewmodel
